requirements.txt
torch
torchvision
flask
opencv-python
pillow
pandas
numpy
tqdm
scikit-learn